+++ lib/src/tabbar_components/tab_item.dart
@@ -6,11 +6,14 @@ import 'notifications_badge.dart';
 /// TabItem is a StatelessWidget used for rendering a single tab in a tab bar.
 /// It supports active/inactive states, labels, notification badges, and custom coloring.
 class TabItem extends StatelessWidget {
-  /// The asset path of the icon when the tab is active.
-  final String activeIcon;
+  /// The icon to display.
+  final Widget? icon;
 
-  /// The asset path of the icon when the tab is inactive.
-  final String inactiveIcon;
+  /// The icon when the tab is active.
+  final Widget? activeIcon;
+
+  /// The icon when the tab is inactive.
+  final Widget? inactiveIcon;
 
   /// The label text of the tab. It's optional and only displayed if [showLabels] is true.
   final String? label;
@@ -25,10 +28,10 @@ class TabItem extends StatelessWidget {
   final bool isFirstActiveIndex;
 
   /// The color to use for the active state of the tab (icon and text).
-  final Color activeColor;
+  final Color? activeColor;
 
   /// The color to use for the inactive state of the tab (icon and text).
-  final Color inactiveColor;
+  final Color? inactiveColor;
 
   /// The color of the notifications badge.
   final Color notificationsBadgeColor;
@@ -39,9 +42,19 @@ class TabItem extends StatelessWidget {
   /// A flag to control whether to show the label under the icon.
   final bool showLabels;
 
+  /// A flag to control whether to override the color of the icons.
+  final bool overrideIconsColor;
+
+  /// The text style to use for the label when the tab is active.
+  final TextStyle? activeLabelTextStyle;
+
+  /// The text style to use for the label when the tab is inactive.
+  final TextStyle? inactiveLabelTextStyle;
+
   /// Constructor for TabItem.
   const TabItem({
     super.key,
+    this.icon,
     required this.activeIcon,
     required this.inactiveIcon,
     required this.label,
@@ -53,6 +66,9 @@ class TabItem extends StatelessWidget {
     required this.notificationsBadgeColor,
     this.showNotifications = false,
     this.showLabels = false,
+    required this.overrideIconsColor,
+    this.activeLabelTextStyle,
+    this.inactiveLabelTextStyle,
   });
 
   @override
@@ -62,12 +78,16 @@ class TabItem extends StatelessWidget {
       children: [
         TabInnerItem(
           active: active,
+          icon: icon,
           activeIcon: activeIcon,
           inactiveIcon: inactiveIcon,
           activeColor: activeColor,
           inactiveColor: inactiveColor,
           label: label,
           showLabels: showLabels,
+          overrideIconsColor: overrideIconsColor,
+          activeLabelTextStyle: activeLabelTextStyle,
+          inactiveLabelTextStyle: inactiveLabelTextStyle,
         ),
         NotificationsBadge(
           showNotifications: showNotifications,